% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_default.R
\name{.is_allFalse}
\alias{.is_allFalse}
\title{Tests if an object is entirely comprised of \code{FALSE}s}
\usage{
.is_allFalse(x)
}
\arguments{
\item{x}{The object to test if is entirely comprised of \code{FALSE} values}
}
\value{
A boolean (\code{TRUE} or \code{FALSE})
}
\description{
This function tests if a passed object is entirely comprised
of \code{FALSE} values.
}
\details{
This function is exported in order to provide one of the default
tests for the \verb{\%|||\%} function and is not really intended for
use outside of that context.
}
\examples{
{
  .is_allFalse(c(FALSE,FALSE,TRUE)) # Will return FALSE
}

}
