% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_default.R
\name{.is_allNA}
\alias{.is_allNA}
\title{Tests if a vector is entirely comprised of NAs}
\usage{
.is_allNA(x)
}
\arguments{
\item{x}{The object to test if is entirely comprised of \code{NA} values}
}
\value{
A boolean (\code{TRUE} or \code{FALSE})
}
\description{
This function tests if a passed object is entirely comprised
of \code{NA} values.
}
\details{
This function is exported in order to provide one of the default
tests for the \verb{\%|||\%} function and is not really intended for
use outside of that context.
}
\examples{
{
  .is_allNA(c(NA,NA,"NA")) # Will return FALSE
}
}
