% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_default.R
\name{.is_length_zero}
\alias{.is_length_zero}
\title{Tests if an object is of length 0}
\usage{
.is_length_zero(x)
}
\arguments{
\item{x}{The object to test if is length(0)}
}
\value{
A boolean (\code{TRUE} or \code{FALSE})
}
\description{
This function tests if a passed object is of length 0.
}
\details{
This function is exported in order to provide one of the default
tests for the \verb{\%|||\%} function and is not really intended for
use outside of that context.
}
\examples{
{
  .is_allFalse(c(FALSE,FALSE,TRUE)) # Will return FALSE
}

}
